Post code review improvement
LocalDate instead of Timestamp.
Use of Mapper convert Java data to SQL data and vice versa.
Use of Object instead of primitive types in Beans.
Integration on Company Bean inside Computer Bean.
Use of Command pattern with CommandFactory to execute command Update_a_computer, List_companies, Create_a_computer, Show_computer_details, List_computers, Delete_a_computer.
Use of own exception on SQL exception instead of to reduce SQL coupling.
Singleton pattern improved to CoManager and added to DAO and Service.

+++ src/org/excilys/test/TestCompany.java
@@ -31,7 +31,7 @@ public class TestCompany {
 	
 	@Test
 	public void testNotEmpty2() {
-		int id = 3;
+		Long id = (long) 3;
 		String name = "test1";
 		
 		company = new Company(id, name);
